docs: Update download links to match latest version (#3375)
* fix: Update download links to match latest version

I'm proposing to change the download reference to the latest releases on Github (URL: BotFramework-Composer/**releases/latest**). 
Currently, the individual download links for Windows, Mac and Linux use an aka.ms reference which is still pointing to version 1.0.0. It would require continuous updating of the aka.ms links to always refer to the latest releases. 
I believe pointing directly to the latest releases on Github would be more manageable in the long run, but feel free to let me know if updating the aka.ms links is still preferred for other reasons.

* Add link to releases/latest

Co-authored-by: Chris Whitten <christopher.whitten@microsoft.com>